requirements.txt
aioredis==1.3.1
altgraph==0.17.2
appdirs==1.4.4
asgiref==3.3.4
astroid==2.4.2
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.6
awsebcli==3.20.0
botocore==1.20.112
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
channels==3.0.4
channels-redis==3.3.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.3.1
django-user-agents==0.4.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
filelock==3.0.12
future==0.16.0
gunicorn==20.0.4
haversine==2.3.1
hiredis==2.0.0
httpie==2.4.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isort==5.6.4
jdcal==1.4.1
jmespath==0.10.0
lazy-object-proxy==1.4.3
lxml==4.7.1
mailer==0.8.1
markdown2==2.4.0
MarkupPy==1.14
mccabe==0.6.1
MouseInfo==0.1.3
mpmath==1.2.1
msgpack==1.0.2
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.20.1
odfpy==1.4.1
opencv-python==4.5.1.48
openpyxl==3.0.5
pandas==1.3.3
pathspec==0.5.9
pbr==5.6.0
pefile==2021.9.3
Pillow==8.2.0
pip-date==1.0.5
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycodestyle==2.7.0
pycparser==2.20
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.9.0
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
pylint==2.6.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyperclip==1.8.1
pypiwin32==223
pypng==0.0.21
PyQRCode==1.2.1
PyRect==0.1.4
PyScreeze==0.1.26
PySocks==1.7.1
python-dateutil==2.8.2
PyTweening==1.0.3
pytz==2020.1
pywin32==301
pywin32-ctypes==0.2.0
PyYAML==5.3.1
qrcode==7.3.1
quick-mailer==0.1.0
requests==2.24.0
requests-toolbelt==0.9.1
selenium==3.141.0
semantic-version==2.8.5
service-identity==21.1.0
six==1.14.0
sqlparse==0.4.1
static3==0.7.0
stevedore==3.3.0
sympy==1.8
tablib==2.0.0
termcolor==1.1.0
toml==0.10.2
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.25.10
user-agents==2.2.0
uuid==1.30
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wcwidth==0.1.9
whitenoise==5.2.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.4.0
